crypto: msm: Make changes to the reaper timer
Change the periodical reaper timer to on-demand basis. The reaper timer
is only set to disable clock after the clock is enabled, and it is
removed when there is no crypto operations. So that we can prevent the
timers from unnecessarily waking up the CPU from power collapse. The
timer interval is changed to 1 second.

CRs-Fixed: 719848
Change-Id: I88901653b779315de558a33ec6d92979c0176e28
Signed-off-by: Zhen Kong <zkong@codeaurora.org>
Signed-off-by: Mekala Natarajan <mekalan@codeaurora.org>
Signed-off-by: placiano <placiano80@gmail.com>